Not using this file yet
As of now, _fullname = "first_last" with no regard to diacriticals.
Since we'll be using http://msrd.io/first_last as a URL to an
individual's home page, we'll have to test the diacriticals.

Otherwise, we can start using this batch file to map UTF-8 names
to simplistic 7-bit ASCII equivalents.