Update rdlist.yaml
I don't know shit about github, except that the UX could stand some improvement. But this is the only kind of editing it lets me do. I guess I'm supposed to submit a "pull request", though I don't know what that does either. But I'll do it and hope for the best.